add clang formatting; reorganize code
