+++ CMakeLists.txt
@@ -49,3 +49,46 @@ target_link_libraries(example2D ${PROJECT_NAME})
 # displaying the header files in IDEs.
 FILE(GLOB_RECURSE LibFiles "include/*.hpp")
 add_custom_target(headers SOURCES ${LibFiles})
+
+#===============================================================================
+# Code Formatting
+#===============================================================================
+if(LRASTAR_VERBOSE)
+  message(STATUS "")
+  message(STATUS "[ Code Formatting ]")
+endif()
+
+find_program(
+  CLANG_FORMAT_EXECUTABLE
+  NAMES clang-format-3.8 clang-format
+)
+
+get_property(formatting_files GLOBAL PROPERTY LRASTAR_FORMAT_FILES)
+list(LENGTH formatting_files formatting_files_length)
+message(STATUS "Formatting on ${formatting_files_length} source files.")
+
+if (CLANG_FORMAT_EXECUTABLE)
+
+  if(LRASTAR_VERBOSE)
+    message(STATUS "Looking for clang-format - ${CLANG_FORMAT_EXECUTABLE} found")
+  endif()
+
+  add_custom_target(format
+                    COMMAND ${CMAKE_COMMAND} -E echo "Formatting code... "
+                    COMMAND ${CLANG_FORMAT_EXECUTABLE} -style=file -i ${formatting_files}
+                    COMMAND ${CMAKE_COMMAND} -E echo "Done."
+                    DEPENDS ${CLANG_FORMAT_EXECUTABLE}
+                    WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/include/LRAstar)
+
+  add_custom_target(check-format
+                    COMMAND ${CMAKE_COMMAND} -E echo "Checking code format... "
+                    COMMAND ${CMAKE_SOURCE_DIR}/tools/check_format.sh ${CLANG_FORMAT_EXECUTABLE} ${formatting_files}
+                    COMMAND ${CMAKE_COMMAND} -E echo "Done."
+                    DEPENDS ${CLANG_FORMAT_EXECUTABLE}
+                    WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/include/LRAstar)
+
+else()
+
+  message(STATUS "Looking for clang-format - NOT found, please install clang-format to enable automatic code formatting")
+
+endif()
