+++ include/LRAstar/LRAstar.hpp
@@ -1,42 +1,7 @@
-/*********************************************************************
-* Software License Agreement (BSD License)
-*
-*  Copyright (c) 2018, University of Washington
-*  All rights reserved.
-*
-*  Redistribution and use in source and binary forms, with or without
-*  modification, are permitted provided that the following conditions
-*  are met:
-*
-*   * Redistributions of source code must retain the above copyright
-*     notice, this list of conditions and the following disclaimer.
-*   * Redistributions in binary form must reproduce the above
-*     copyright notice, this list of conditions and the following
-*     disclaimer in the documentation and/or other materials provided
-*     with the distribution.
-*   * Neither the name of the University of Washington nor the names
-*     of its contributors may be used to endorse or promote products
-*     derived from this software without specific prior written
-*     permission.
-*
-*  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-*  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-*  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-*  FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-*  COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-*  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-*  BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-*  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-*  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
-*  LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
-*  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-*  POSSIBILITY OF SUCH DAMAGE.
-*********************************************************************/
-
 /* Authors: Aditya Vamsikrishna Mandalika */
 
-#ifndef LRASTAR_HPP_
-#define LRASTAR_HPP_
+#ifndef LRASTAR_LRASTAR_HPP_
+#define LRASTAR_LRASTAR_HPP_
 
 // STL headers
 #include <vector>
@@ -124,10 +89,7 @@ public:
     /// Children
     std::vector<std::size_t> children;
 
-    /// Flag to check if vertex is within the lazyband
-    bool inLazyBand;
-
-    /// Visited TODO(avk): Needed?
+    /// Visited
     bool visited;
 
     /// Collision status
@@ -400,4 +362,4 @@ private:
 }; // class LRAstar
 
 } // namespace LRAstar
-#endif // LRASTAR_HPP_
+#endif // LRASTAR_LRASTAR_HPP_
