+++ include/LRAstar/util.hpp
